requirements.txt
# Core dependencies
fastapi
uvicorn
websockets
pydantic
python-dotenv
openai
twilio